move 'reject' tests to separate file
