+++ types/ramda/tsconfig.json
@@ -189,6 +189,7 @@
         "test/reduce-tests.ts",
         "test/reduceBy-tests.ts",
         "test/reduceRight-tests.ts",
-        "test/reduceWhile-tests.ts"
+        "test/reduceWhile-tests.ts",
+        "test/reject-tests.ts"
     ]
 }
